Fix formatting in MBeanWrapper, move JMX to its own class
