+++ test/distributed/org/apache/cassandra/distributed/impl/AbstractCluster.java
@@ -187,7 +187,7 @@ public abstract class AbstractCluster<I extends IInstance> implements ICluster<I
 
         {
             // Indicate that we are running in the in-jvm dtest environment
-            CassandraRelevantProperties.IS_IN_JVM_DTEST.setBoolean(true);
+            CassandraRelevantProperties.TEST_IS_IN_JVM_DTEST.setBoolean(true);
             // those properties may be set for unit-test optimizations; those should not be used when running dtests
             CassandraRelevantProperties.TEST_FLUSH_LOCAL_SCHEMA_CHANGES.reset();
             CassandraRelevantProperties.NON_GRACEFUL_SHUTDOWN.reset();
