+++ test/distributed/org/apache/cassandra/distributed/impl/Instance.java
@@ -21,19 +21,14 @@ package org.apache.cassandra.distributed.impl;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
-import java.lang.reflect.Field;
 import java.net.BindException;
-import java.net.InetAddress;
 import java.net.InetSocketAddress;
-import java.net.MalformedURLException;
 import java.nio.ByteBuffer;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.security.Permission;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -48,12 +43,6 @@ import java.util.stream.Stream;
 import javax.management.ListenerNotFoundException;
 import javax.management.Notification;
 import javax.management.NotificationListener;
-import javax.management.remote.JMXConnector;
-import javax.management.remote.JMXConnectorFactory;
-import javax.management.remote.JMXConnectorServer;
-import javax.management.remote.JMXServiceURL;
-import javax.management.remote.rmi.RMIConnectorServer;
-import javax.management.remote.rmi.RMIJRMPServerImpl;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.util.concurrent.Uninterruptibles;
@@ -103,7 +92,6 @@ import org.apache.cassandra.distributed.mock.nodetool.InternalNodeProbe;
 import org.apache.cassandra.distributed.mock.nodetool.InternalNodeProbeFactory;
 import org.apache.cassandra.distributed.shared.Metrics;
 import org.apache.cassandra.distributed.shared.ThrowingRunnable;
-import org.apache.cassandra.utils.ReflectionUtils;
 import org.apache.cassandra.gms.Gossiper;
 import org.apache.cassandra.hints.DTestSerializer;
 import org.apache.cassandra.hints.HintsService;
@@ -156,25 +144,18 @@ import org.apache.cassandra.utils.Closeable;
 import org.apache.cassandra.utils.DiagnosticSnapshotService;
 import org.apache.cassandra.utils.ExecutorUtils;
 import org.apache.cassandra.utils.FBUtilities;
-import org.apache.cassandra.utils.JMXServerUtils;
 import org.apache.cassandra.utils.JVMStabilityInspector;
-import org.apache.cassandra.utils.MBeanWrapper;
-import org.apache.cassandra.utils.RMIClientSocketFactoryImpl;
 import org.apache.cassandra.utils.Throwables;
 import org.apache.cassandra.utils.concurrent.Ref;
 import org.apache.cassandra.utils.logging.LoggingSupportFactory;
 import org.apache.cassandra.utils.memory.BufferPools;
 import org.apache.cassandra.utils.progress.jmx.JMXBroadcastExecutor;
-import sun.rmi.transport.tcp.TCPEndpoint;
 
 import static java.util.concurrent.TimeUnit.MINUTES;
 import static org.apache.cassandra.concurrent.ExecutorFactory.Global.executorFactory;
 import static org.apache.cassandra.config.CassandraRelevantProperties.CONSISTENT_RANGE_MOVEMENT;
 import static org.apache.cassandra.config.CassandraRelevantProperties.CONSISTENT_SIMULTANEOUS_MOVES_ALLOW;
-import static org.apache.cassandra.config.CassandraRelevantProperties.JAVA_RMI_DGC_LEASE_VALUE_IN_JVM_DTEST;
-import static org.apache.cassandra.config.CassandraRelevantProperties.ORG_APACHE_CASSANDRA_DISABLE_MBEAN_REGISTRATION;
 import static org.apache.cassandra.config.CassandraRelevantProperties.RING_DELAY;
-import static org.apache.cassandra.config.CassandraRelevantProperties.SUN_RMI_TRANSPORT_TCP_THREADKEEPALIVETIME;
 import static org.apache.cassandra.config.CassandraRelevantProperties.TEST_CASSANDRA_SUITENAME;
 import static org.apache.cassandra.config.CassandraRelevantProperties.TEST_CASSANDRA_TESTTAG;
 import static org.apache.cassandra.distributed.api.Feature.BLANK_GOSSIP;
@@ -192,18 +173,12 @@ import static org.apache.cassandra.utils.Clock.Global.nanoTime;
  */
 public class Instance extends IsolatedExecutor implements IInvokableInstance
 {
-    private static final int RMI_KEEPALIVE_TIME = 1000;
     private Logger inInstancelogger; // Defer creation until running in the instance context
     public final IInstanceConfig config;
     private volatile boolean initialized = false;
     private volatile boolean internodeMessagingStarted = false;
     private final AtomicLong startedAt = new AtomicLong();
-    private JMXConnectorServer jmxConnectorServer;
-    private JMXServerUtils.JmxRegistry registry;
-    private RMIJRMPServerImpl jmxRmiServer;
-    private MBeanWrapper.InstanceMBeanWrapper wrapper;
-    private RMIClientSocketFactoryImpl clientSocketFactory;
-    private CollectingRMIServerSocketFactoryImpl serverSocketFactory;
+    private IsolatedJmx isolatedJmx;
 
     @Deprecated
     Instance(IInstanceConfig config, ClassLoader classLoader)
@@ -789,168 +764,15 @@ public class Instance extends IsolatedExecutor implements IInvokableInstance
         initialized = true;
     }
 
-    private void startJmx() throws InterruptedException
+    private void startJmx()
     {
-        try
-        {
-            // Several RMI threads hold references to in-jvm dtest objects, and are, by default, kept
-            // alive for long enough (minutes) to keep classloaders from being collected.
-            // Set these two system properties to a low value to allow cleanup to occur fast enough
-            // for GC to collect our classloaders.
-            JAVA_RMI_DGC_LEASE_VALUE_IN_JVM_DTEST.setInt(RMI_KEEPALIVE_TIME);
-            SUN_RMI_TRANSPORT_TCP_THREADKEEPALIVETIME.setInt(RMI_KEEPALIVE_TIME);
-            ORG_APACHE_CASSANDRA_DISABLE_MBEAN_REGISTRATION.setBoolean(false);
-            InetAddress addr = config.broadcastAddress().getAddress();
-
-            int jmxPort = config.jmxPort();
-
-            String hostname = addr.getHostAddress();
-            wrapper = new MBeanWrapper.InstanceMBeanWrapper(hostname + ":" + jmxPort);
-            ((MBeanWrapper.DelegatingMbeanWrapper) MBeanWrapper.instance).setDelegate(wrapper);
-            Map<String, Object> env = new HashMap<>();
-
-            serverSocketFactory = new CollectingRMIServerSocketFactoryImpl(addr);
-            env.put(RMIConnectorServer.RMI_SERVER_SOCKET_FACTORY_ATTRIBUTE,
-                    serverSocketFactory);
-            clientSocketFactory = new RMIClientSocketFactoryImpl(addr);
-            env.put(RMIConnectorServer.RMI_CLIENT_SOCKET_FACTORY_ATTRIBUTE,
-                    clientSocketFactory);
-
-            // configure the RMI registry
-            registry = new JMXServerUtils.JmxRegistry(jmxPort,
-                                                      clientSocketFactory,
-                                                      serverSocketFactory,
-                                                      "jmxrmi");
-
-            // Mark the JMX server as a permanently exported object. This allows the JVM to exit with the
-            // server running and also exempts it from the distributed GC scheduler which otherwise would
-            // potentially attempt a full GC every `sun.rmi.dgc.server.gcInterval` millis (default is 3600000ms)
-            // For more background see:
-            //   - CASSANDRA-2967
-            //   - https://www.jclarity.com/2015/01/27/rmi-system-gc-unplugged/
-            //   - https://bugs.openjdk.java.net/browse/JDK-6760712
-            env.put("jmx.remote.x.daemon", "true");
-
-            // Set the port used to create subsequent connections to exported objects over RMI. This simplifies
-            // configuration in firewalled environments, but it can't be used in conjuction with SSL sockets.
-            // See: CASSANDRA-7087
-            int rmiPort = config.jmxPort();
-
-            // We create the underlying RMIJRMPServerImpl so that we can manually bind it to the registry,
-            // rather then specifying a binding address in the JMXServiceURL and letting it be done automatically
-            // when the server is started. The reason for this is that if the registry is configured with SSL
-            // sockets, the JMXConnectorServer acts as its client during the binding which means it needs to
-            // have a truststore configured which contains the registry's certificate. Manually binding removes
-            // this problem.
-            // See CASSANDRA-12109.
-            jmxRmiServer = new RMIJRMPServerImpl(rmiPort,clientSocketFactory, serverSocketFactory,
-                                                 env);
-            JMXServiceURL serviceURL = new JMXServiceURL("rmi", hostname, rmiPort);
-            jmxConnectorServer = new RMIConnectorServer(serviceURL, env, jmxRmiServer, wrapper.getMbs());
-
-            jmxConnectorServer.start();
-
-            registry.setRemoteServerStub(jmxRmiServer.toStub());
-            JMXServerUtils.logJmxServiceUrl(addr, jmxPort);
-            waitForJmxAvailability(hostname, jmxPort, env);
-        }
-        catch (IOException e)
-        {
-            throw new RuntimeException("Feature.JMX was enabled but could not be started.", e);
-        }
+        this.isolatedJmx = new IsolatedJmx(this, inInstancelogger);
+        isolatedJmx.startJmx();
     }
 
-    private void waitForJmxAvailability(String hostname, int rmiPort, Map<String, Object> env) throws InterruptedException, MalformedURLException
+    private void stopJmx() throws NoSuchFieldException, InterruptedException, IllegalAccessException
     {
-        String url = String.format("service:jmx:rmi:///jndi/rmi://%s:%d/jmxrmi", hostname, rmiPort);
-        JMXServiceURL serviceURL = new JMXServiceURL(url);
-        int attempts = 0;
-        Throwable lastThrown = null;
-        while (attempts < 20)
-        {
-            attempts++;
-            try (JMXConnector ignored = JMXConnectorFactory.connect(serviceURL, env))
-            {
-                inInstancelogger.info("Connected to JMX server at {} after {} attempt(s)",
-                                      url, attempts);
-                return;
-            }
-            catch (MalformedURLException e)
-            {
-                throw new RuntimeException(e);
-            }
-            catch (Throwable thrown)
-            {
-                lastThrown = thrown;
-            }
-            inInstancelogger.info("Could not connect to JMX on {} after {} attempts. Will retry.", url, attempts);
-            Thread.sleep(1000);
-        }
-        throw new RuntimeException("Could not start JMX - unreachable after 20 attempts", lastThrown);
-    }
-
-    private void stopJmx() throws IllegalAccessException, NoSuchFieldException, InterruptedException
-    {
-        if (!config.has(JMX))
-            return;
-        // First, swap the mbean wrapper back to a NoOp wrapper
-        // This prevents later attempts to unregister mbeans from failing in Cassandra code, as we're going to
-        // unregister all of them here
-        ((MBeanWrapper.DelegatingMbeanWrapper) MBeanWrapper.instance).setDelegate(new MBeanWrapper.NoOpMBeanWrapper());
-        try
-        {
-            wrapper.close();
-        }
-        catch (Throwable e)
-        {
-            inInstancelogger.warn("failed to close wrapper.", e);
-        }
-        try
-        {
-            jmxConnectorServer.stop();
-        }
-        catch (Throwable e)
-        {
-            inInstancelogger.warn("failed to close jmxConnectorServer.", e);
-        }
-        try
-        {
-            registry.close();
-        }
-        catch (Throwable e)
-        {
-            inInstancelogger.warn("failed to close registry.", e);
-        }
-        try
-        {
-            serverSocketFactory.close();
-        }
-        catch (Throwable e)
-        {
-            inInstancelogger.warn("failed to close serverSocketFactory.", e);
-        }
-        // The TCPEndpoint class holds references to a class in the in-jvm dtest framework
-        // which transitively has a reference to the InstanceClassLoader, so we need to
-        // make sure to remove the reference to them when the instance is shutting down
-        clearMapField(TCPEndpoint.class, null, "localEndpoints");
-        Thread.sleep(2 * RMI_KEEPALIVE_TIME); // Double the keep-alive time to give Distributed GC some time to clean up
-    }
-
-    private <K, V> void clearMapField(Class<?> clazz, Object instance, String mapName)
-    throws IllegalAccessException, NoSuchFieldException {
-        Field mapField = ReflectionUtils.getField(clazz, mapName);
-        mapField.setAccessible(true);
-        Map<K, V> map = (Map<K, V>) mapField.get(instance);
-        // Because multiple instances can be shutting down at once,
-        // synchronize on the map to avoid ConcurrentModificationException
-        synchronized (map)
-        {
-            for (Iterator<Map.Entry<K, V>> it = map.entrySet().iterator(); it.hasNext(); )
-            {
-                it.next();
-                it.remove();
-            }
-        }
+        isolatedJmx.stopJmx();
     }
 
     // Update the messaging versions for all instances
